refactor: added chunk id reference for sources
